+++ ng2Client/app/modules/data/entities/user-element-field.ts
@@ -6,4 +6,8 @@ export class UserElementField extends EntityBase {
     UserId: number = 0;
     UserElementFieldId: number = 0;
     Rating: number = 0;
+
+    static initializer(entity: UserElementField) {
+        super.initializer(entity);
+    }
 }