+++ ng2Client/app/modules/data/entities/user-resource-pool.ts
@@ -6,4 +6,8 @@ export class UserResourcePool extends EntityBase {
     UserId: number = 0;
     ResourcePoolId: number = 0;
     ResourcePoolRate: number = 0;
+
+    static initializer(entity: UserResourcePool) {
+        super.initializer(entity);
+    }
 }